[ticket/15960] Add SQL transactions to functions_admin.php
PHPBB3-15960

And simultaneously add a check-in and count before to run the queries.

+++ phpBB/includes/functions_admin.php
@@ -3042,14 +3042,21 @@ function tidy_database()
 	}
 	$db->sql_freeresult($result);
 
-	// Delete those rows from the acl tables not having listed the forums above
-	$sql = 'DELETE FROM ' . ACL_GROUPS_TABLE . '
-		WHERE ' . $db->sql_in_set('forum_id', $forum_ids, true);
-	$db->sql_query($sql);
+	if (count($forum_ids))
+	{
+		$db->sql_transaction('begin');
 
-	$sql = 'DELETE FROM ' . ACL_USERS_TABLE . '
-		WHERE ' . $db->sql_in_set('forum_id', $forum_ids, true);
-	$db->sql_query($sql);
+		// Delete those rows from the acl tables not having listed the forums above
+		$sql = 'DELETE FROM ' . ACL_GROUPS_TABLE . '
+			WHERE ' . $db->sql_in_set('forum_id', $forum_ids, true);
+		$db->sql_query($sql);
+
+		$sql = 'DELETE FROM ' . ACL_USERS_TABLE . '
+			WHERE ' . $db->sql_in_set('forum_id', $forum_ids, true);
+		$db->sql_query($sql);
+
+		$db->sql_transaction('commit');
+	}
 
 	$config->set('database_last_gc', time(), false);
 }
